requirements.txt
Flask==3.0.0
requests==2.28.1
pillow
flask_sqlalchemy

